requirements.txt
youtube-transcript-api
google-generativeai
python-dotenv
aiogram==3.13.1
markdown
